Feat: 중간점검미완성

+++ src/main/java/com/sparta/spartaboardtest/service/BoardService.java
@@ -2,6 +2,7 @@ package com.sparta.spartaboardtest.service;
 
 
 import com.sparta.spartaboardtest.dto.BoardRequestDto;
+import com.sparta.spartaboardtest.dto.BoardResponseDto;
 import com.sparta.spartaboardtest.entity.Board;
 import com.sparta.spartaboardtest.repository.BoardRepository;
 import lombok.RequiredArgsConstructor;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -28,17 +30,20 @@ public class BoardService {
     }
 
     @Transactional
-    public Long update(Long id, BoardRequestDto requestDto) {
+    public Long update(Long id, BoardResponseDto responseDto) {
         Board board = boardRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
         );
-        board.update(requestDto);
+        board.update(responseDto);
         return board.getId();
     }
 
     @Transactional
-    public Long deleteBoard(Long id) {
-        boardRepository.deleteById(id);
+    public Long deleteBoard(Long id, String password) {
+        BoardRequestDto requestDto = new BoardRequestDto();
+        if (Objects.equals(password, requestDto.getPassword())) {
+            boardRepository.deleteById(id);
+        }
         return id;
     }
 }
